Improve UX and add instructions
