+++ src/components/Button.tsx
@@ -1,5 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
+import { clickableText } from "../constants/compositeStyles";
+import { click } from "@testing-library/user-event/dist/click";
 
 export enum ButtonColors {
   Blue = "#3F8CFF",
@@ -17,6 +19,11 @@ interface ButtonBaseProps {
 interface ButtonLinkProps extends ButtonBaseProps {
   href: string;
   external?: boolean;
+}
+
+interface InlineButtonProps {
+  text: string;
+  onClick?: () => void;
 }
 
 function ButtonBase(props: ButtonBaseProps) {
@@ -46,6 +53,13 @@ export function ButtonLink(props: ButtonLinkProps) {
       <ButtonBase {...rest} />
     </Link>
   );
+}
+
+export function InlineButton(props: InlineButtonProps) {
+  const { onClick, text } = props
+  return (
+    <span className={onClick ? clickableText : undefined} onClick={onClick} role='button'>{text}</span>
+  )
 }
 
 export default function Button(props: ButtonBaseProps) {
