+++ src/components/Card.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 
-export default function Card(props: { children: React.ReactNode }) {
-  const { children } = props;
+export default function Card({ children, className = ''}: { children: React.ReactNode, className?: string }) {
   return (
-    <div className="max-w-xl px-6 mx-auto first:mt-20 mt-12">{children}</div>
+    <div className={`max-w-xl px-6 mx-auto first:mt-20 mt-12 ${className}`}>{children}</div>
   );
 }
