+++ src/constants/compositeStyles.tsx
@@ -1,1 +1,1 @@
-export const clickableText = "underline underline-offset-2";
+export const clickableText = "cursor-pointer text-blue-600 hover:text-blue-700 underline underline-offset-2";
